add productsRequsts to heroku
