+++ CleanCode/App/Entity/Product/GetProductResult.swift
@@ -8,15 +8,5 @@
 import Foundation
 
 struct GetProductResult: Codable {
-    let result: Int
-    let productName: String
-    let productPrice: Int
-    let productDescription: String
-    
-    enum CodingKeys: String, CodingKey {
-        case result = "result"
-        case productName = "product_name"
-        case productPrice = "product_price"
-        case productDescription = "product_description"
-    }
+    let product: Product?
 }
