+++ CleanCode/Core/Extensions/Products/ProductRequest.swift
@@ -10,8 +10,8 @@ import Alamofire
 
 struct ProductRequest: RequestRouter {
     let baseUrl: URL
-    let method: HTTPMethod = .get
-    let path: String = "getGoodById.json"
+    let method: HTTPMethod = .post
+    let path: String = "getProduct"
     let productID: Int
     var parameters: Parameters? {
         return [
