+++ CleanCode/Core/Extensions/Products/ProductsRequest.swift
@@ -10,10 +10,10 @@ import Alamofire
 
 struct ProductsRequest: RequestRouter {
     let baseUrl: URL
-    let method: HTTPMethod = .get
-    let path: String = "catalogData.json"
-    let pageNumber: Int
-    let categoryID: Int
+    let method: HTTPMethod = .post
+    let path: String = "getProducts"
+    let pageNumber: String
+    let categoryID: String
     var parameters: Parameters? {
         return [
             "page_number": pageNumber,
